react/requirements.txt
asgiref==3.2.7
Django==3.0.4
django-cors-headers==3.2.1
django-taggit==1.2.0
djangorestframework==3.11.0
gunicorn==19.7.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
whitenoise==3.3.1
